refactor: extract budgetCalculation module from calculator (Phase 4a.6)
Move skill/trade/language/ancestry budget calculations into
calculatorModules/budgetCalculation.ts. All snapshot tests pass
with identical output.
